@@ -28,44 +28,45 @@
   "license": "MIT",
   "schematics": "./src/collection.json",
   "dependencies": {
-    "@angular-devkit/core": "^18.2.8",
-    "@angular-devkit/schematics": "^18.2.8",
-    "axios": "^1.12.0",
-    "lodash": "^4.17.21",
+    "@angular-devkit/core": "^20.3.15",
+    "axios": "^1.13.2",
+    "lodash": "^4.17.23",
     "pluralize": "^8.0.0",
     "rxjs": "^7.8.1"
   },
   "devDependencies": {
-    "@angular-devkit/schematics-cli": "^20.1.6",
-    "@angular/compiler": "^18.2.8",
-    "@angular/core": "^18.2.8",
-    "@eslint/js": "^9.8.0",
-    "@types/jest": "^29.5.13",
-    "@types/lodash": "^4.17.10",
-    "@types/node": "^20.16.11",
+    "@angular-devkit/schematics": "^20.3.15",
+    "@angular-devkit/schematics-cli": "^21.1.1",
+    "@angular/compiler": "^21.1.1",
+    "@angular/core": "^21.1.1",
+    "@eslint/js": "^9.39.2",
+    "@jest/globals": "^30.2.0",
+    "@types/lodash": "^4.17.23",
+    "@types/node": "^22.19.7",
     "@types/pluralize": "^0.0.33",
-    "@typescript-eslint/eslint-plugin": "^8.8.1",
+    "@typescript-eslint/eslint-plugin": "^8.53.1",
     "@typescript-eslint/parser": "^8.8.1",
     "@typescript-eslint/typescript-estree": "^8.8.1",
-    "baseline-browser-mapping": "^2.9.14",
-    "eslint": "^9.12.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-angular": "^4.1.0",
-    "eslint-plugin-jest": "^28.8.3",
-    "eslint-plugin-jest-dom": "^5.4.0",
+    "eslint": "^9.39.2",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-angular": "^5.0.0",
+    "eslint-plugin-jest": "^29.12.1",
+    "eslint-plugin-jest-dom": "^5.5.0",
     "eslint-plugin-jest-formatting": "^3.1.0",
+    "eslint-plugin-ngrx": "^2.1.4",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
+    "eslint-plugin-prettier": "^5.5.5",
     "eslint-plugin-simple-import-sort": "^12.1.1",
-    "eslint-plugin-sonarjs": "^2.0.3",
+    "eslint-plugin-sonarjs": "^3.0.5",
     "eslint-plugin-standard": "^4.1.0",
     "eslint-plugin-typescript": "^0.14.0",
-    "jest": "^29.7.0",
+    "jest": "^30.2.0",
     "jest-junit": "^16.0.0",
+    "prettier": "3.8.1",
     "schematics": "^0.2.0",
-    "snyk": "^1.1297.3",
-    "ts-jest": "^29.2.5",
-    "typescript": "^5.6.3",
-    "typescript-eslint": "^8.8.1"
+    "snyk": "^1.1302.1",
+    "ts-jest": "^29.4.6",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.53.1"
   }
 }
