@@ -1,12 +1,13 @@
+import { describe, expect, test } from '@jest/globals';
 import { Tree } from '@angular-devkit/schematics';
 import {
   SchematicTestRunner,
   UnitTestTree,
 } from '@angular-devkit/schematics/testing';
-import * as path from 'path';
+import * as path from 'node:path';
 import { IOptions } from '../shared';
 import { classify } from '@angular-devkit/core/src/utils/strings';
-import * as pluralize from 'pluralize';
+import { plural } from 'pluralize';
 
 const collectionPath = path.join(__dirname, '../collection.json');
 
@@ -25,26 +26,26 @@ describe('imng-list', () => {
     const tree: UnitTestTree = await runner.runSchematic('imng-list', options, Tree.empty());
 
     expect(tree.files).toEqual([
-      `/test/${pluralize(options.name)}-list/index.ts`,
-      `/test/${pluralize(options.name)}-list/list.component.html`,
-      `/test/${pluralize(options.name)}-list/list.component.scss`,
-      `/test/${pluralize(options.name)}-list/list.component.spec.ts`,
-      `/test/${pluralize(options.name)}-list/list.component.ts`,
-      `/test/${pluralize(options.name)}-list/list.facade.spec.ts`,
-      `/test/${pluralize(options.name)}-list/list.facade.ts`,
-      `/test/${pluralize(options.name)}-list/list.grid-state.ts`,
+      `/test/${plural(options.name)}-list/index.ts`,
+      `/test/${plural(options.name)}-list/list.component.html`,
+      `/test/${plural(options.name)}-list/list.component.scss`,
+      `/test/${plural(options.name)}-list/list.component.spec.ts`,
+      `/test/${plural(options.name)}-list/list.component.ts`,
+      `/test/${plural(options.name)}-list/list.facade.spec.ts`,
+      `/test/${plural(options.name)}-list/list.facade.ts`,
+      `/test/${plural(options.name)}-list/list.grid-state.ts`,
     ]);
 
     const htmlFile = tree.get(
-      `/test/${pluralize(options.name)}-list/list.component.html`
+      `/test/${plural(options.name)}-list/list.component.html`
     );
     let content = htmlFile?.content.toString();
     expect(content).toContain('[field]="props.ADDRESS_LINE_1"');
     expect(content).toContain(`<${options.appPrefix}-${options.name}-add `);
     expect(content).toContain(`<${options.appPrefix}-${options.name}-edit `);
-    expect(content).toMatchSnapshot("list.component.html");
+    expect(content).toMatchSnapshot('list.component.html');
     const gridStateFile = tree.get(
-      `/test/${pluralize(options.name)}-list/list.grid-state.ts`
+      `/test/${plural(options.name)}-list/list.grid-state.ts`
     );
     content = gridStateFile?.content.toString();
     expect(content).toContain(
@@ -53,13 +54,13 @@ describe('imng-list', () => {
     expect(content).toMatchSnapshot();
 
     const componentFile = tree.get(
-      `/test/${pluralize(options.name)}-list/list.component.ts`
+      `/test/${plural(options.name)}-list/list.component.ts`
     );
     content = componentFile?.content.toString();
     expect(content).toContain(`'${options.appPrefix}-${options.name}-list'`);
 
     const facadeSpecFile = tree.get(
-      `/test/${pluralize(options.name)}-list/list.facade.spec.ts`
+      `/test/${plural(options.name)}-list/list.facade.spec.ts`
     );
     content = facadeSpecFile?.content.toString();
     expect(content).toContain(
